+++ Hardware/Motor.h
@@ -2,8 +2,8 @@
 #define __MOTOR_H
 
 void Motor_Init(void);
-void Motor_SetSpeedA(int8_t Speed);
-void Motor_SetSpeedB(int8_t Speed);
-void Motor_SetSpeed(int MotorStatus,int8_t Speed);
+void Motor_SetSpeedA(int16_t Speed);
+void Motor_SetSpeedB(int16_t Speed);
+void Motor_SetSpeed(int MotorStatus,int16_t Speed);
 
 #endif
